Filtering is added to execution engine callback service and removed from service which returns execution result to user(#1046)
