+++ src/main/java/org/ohdsi/webapi/executionengine/service/AnalysisResultFileContentSensitiveInfoService.java
@@ -1,0 +1,7 @@
+package org.ohdsi.webapi.executionengine.service;
+
+import org.ohdsi.webapi.common.sensitiveinfo.SensitiveInfoService;
+import org.ohdsi.webapi.executionengine.entity.AnalysisResultFileContent;
+
+public interface AnalysisResultFileContentSensitiveInfoService extends SensitiveInfoService<AnalysisResultFileContent> {
+}
