--- src/main/java/org/ohdsi/webapi/executionengine/service/AnalysisResultFileSensitiveInfoService.java
@@ -1,7 +1,0 @@
-package org.ohdsi.webapi.executionengine.service;
-
-import org.ohdsi.webapi.common.sensitiveinfo.SensitiveInfoService;
-import org.ohdsi.webapi.executionengine.entity.AnalysisResultFile;
-
-public interface AnalysisResultFileSensitiveInfoService extends SensitiveInfoService<AnalysisResultFile> {
-}
